requirements.txt
alembic==1.14.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
